+++ src/main/java/net/larsmans/infinitybuttons/block/custom/StickyCopperButton.java
@@ -120,7 +120,7 @@ public class StickyCopperButton extends WallMountedBlock {
     }
 
     protected void playClickSound(@Nullable PlayerEntity player, WorldAccess world, BlockPos pos, boolean powered) {
-        world.playSound(powered ? player : null, pos, this.getClickSound(powered), SoundCategory.BLOCKS, 0.3f, powered ? 0.6f : 0.5f);
+        world.playSound(powered ? player : null, pos, this.getClickSound(powered), SoundCategory.BLOCKS, 1f, powered ? 0.6f : 0.5f);
     }
 
     protected SoundEvent getClickSound(boolean var1){