+++ src/main/java/net/larsmans/infinitybuttons/block/custom/StickyCopperLargeButton.java
@@ -123,7 +123,7 @@ public class StickyCopperLargeButton extends WallMountedBlock {
     }
 
     protected void playClickSound(@Nullable PlayerEntity player, WorldAccess world, BlockPos pos, boolean pressed) {
-        world.playSound(pressed ? player : null, pos, this.getClickSound(), SoundCategory.BLOCKS, 0.3f, pressed ? 0.6f : 0.5f);
+        world.playSound(pressed ? player : null, pos, this.getClickSound(), SoundCategory.BLOCKS, 0.3f, pressed ? 1f : 0.5f);
     }
 
     protected SoundEvent getClickSound() {